+++ CataloguesAnywhere/ModEntry.cs
@@ -2,6 +2,7 @@ using CataloguesAnywhere.Framework;
 using ChroniclerCherry.Common.Integrations.GenericModConfigMenu;
 using StardewModdingAPI;
 using StardewModdingAPI.Events;
+using StardewModdingAPI.Utilities;
 using StardewValley;
 
 namespace CataloguesAnywhere;
@@ -37,7 +38,7 @@ internal class ModEntry : Mod
     private void OnGameLaunched(object? sender, GameLaunchedEventArgs e)
     {
         this.AddGenericModConfigMenu(
-            new GenericModConfigMenuIntegrationForCataloguesAnywhere(),
+            new GenericModConfigMenuIntegrationForCataloguesAnywhere(() => this.Config),
             get: () => this.Config,
             set: config => this.Config = config
         );
@@ -49,9 +50,14 @@ internal class ModEntry : Mod
         if (!Context.CanPlayerMove || !this.Config.Enabled)
             return;
 
-        if (this.Config.FurnitureKey.JustPressed())
-            Utility.TryOpenShopMenu("Furniture Catalogue", null as string);
-        else if (this.Config.WallpaperKey.JustPressed())
-            Utility.TryOpenShopMenu("Catalogue", null as string);
+        foreach ((string shopId, KeybindList keybind) in this.Config.Catalogues)
+        {
+            if (keybind.JustPressed())
+            {
+                this.Monitor.Log($"Opening shop ID '{shopId}' per keybind '{keybind}'.");
+                Utility.TryOpenShopMenu(shopId, null as string);
+                break;
+            }
+        }
     }
 }